chore: moving tests to local __tests__ dir
