+++ src/client/utils/__tests__/SearchSourceFactory.test.ts
@@ -1,15 +1,15 @@
 import lunr from "lunr";
-import { SearchDocument } from "../../types";
-import { SearchSourceFactory } from "./SearchSourceFactory";
+import { SearchDocument } from "../../../types";
+import { SearchSourceFactory } from "../SearchSourceFactory";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 require("lunr-languages/lunr.stemmer.support")(lunr);
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-require("../../shared/lunrLanguageZh").lunrLanguageZh(lunr);
+require("../../../shared/lunrLanguageZh").lunrLanguageZh(lunr);
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 require("lunr-languages/lunr.multi")(lunr);
 
-jest.mock("./proxiedGenerated");
+jest.mock("../proxiedGenerated");
 
 describe("SearchSourceFactory", () => {
   const documentsOfTitles: SearchDocument[] = [
