+++ src/client/utils/__tests__/highlightStemmed.test.ts
@@ -1,7 +1,11 @@
-import { MetadataPosition, ChunkIndexRef, HighlightChunk } from "../../types";
-import { highlightStemmed, splitIntoChunks } from "./highlightStemmed";
+import {
+  MetadataPosition,
+  ChunkIndexRef,
+  HighlightChunk,
+} from "../../../types";
+import { highlightStemmed, splitIntoChunks } from "../highlightStemmed";
 
-jest.mock("./proxiedGenerated");
+jest.mock("../proxiedGenerated");
 
 describe("highlightStemmed", () => {
   test.each<[string, MetadataPosition[], string[], number | undefined, string]>(
